fix(study): widen the day's word counter to 64 bits
The per-day word counter is unlike the event counters beside it. They move by
one; this one moves by however many words the caller's own anchor claims were
covered, on the one mutation deliberately left unthrottled. Alternating between
two ends of a long book credits roughly half a book per round trip, so a few
thousand requests reach the 32-bit ceiling.

What lands there is worse than a wrong figure: the overflow aborts the upsert,
and the position write shares that transaction, so the reader would stop being
able to save their place for the rest of that day. The revision is unreleased,
so the column is widened in place rather than by a follow-up; the downgrade
still drops it, and the migration test now round-trips a value past the old
ceiling.

+++ backend/app/infrastructure/db/metadata.py
@@ -722,6 +722,9 @@ study_days = Table(
     Column("reviews_count", Integer, nullable=False, server_default="0"),
     Column("reading_updates", Integer, nullable=False, server_default="0"),
     # Words, not pages: a day's small advances accumulate losslessly and the page
-    # quantum stays a read-time presentation constant (AD-183).
-    Column("words_advanced", Integer, nullable=False, server_default="0"),
+    # quantum stays a read-time presentation constant (AD-183). 64-bit unlike its
+    # neighbours: they count events (+1 each), while this one adds however many words the
+    # caller's own anchor claims, on an endpoint with no rate limit — so an overflow is
+    # reachable, and it would abort the position write sharing the same transaction.
+    Column("words_advanced", BigInteger, nullable=False, server_default="0"),
 )
